Load notification count when resumed
When using the iOS PWA the service worker is not able to send a message
to the client when a notification is clicked (see last commit).
Therefore the notification badge is not updated when a notification is
received.
As workaround when focus is lost and regained the notification count is updated manually.