Improve resume from background on iOS
When resuming the app from background the click on the homescreen icon
seems to trigger a touchStart event.
This triggers a navigation drawer opening when the navigation drawer
triggering area is touched.

So when resuming the application the first click is ignored to not open
the navigation drawer on click.
Additionally when the navigation drawer was open all variables are
reset, otherwise some unwanted effects are triggered, like the overlay
is dismissed but the navigation drawer is still visible.